Completely redesign Inertia CLI structure
