+++ .scripts/release.sh
@@ -13,5 +13,5 @@ make daemon RELEASE="$RELEASE"
 
 # Build Inertia Go binaries for specified platforms
 gox -output="inertia.$(git describe --tags).{{.OS}}.{{.Arch}}" \
-    -ldflags "-w -s -X github.com/ubclaunchpad/inertia/cmd.Version=$RELEASE" \
+    -ldflags "-w -s -X main.Version=$RELEASE" \
     -osarch="$PLATFORMS" \
